# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "description": "API documentation for Laravel + React project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get authenticated user profile",
                "operationId": "39af45cbadb8ce9e13521407f440ff2d",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reset-password": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Reset user password",
                "operationId": "fb2894f3487b382b2e1e4e5d639eba56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Password reset successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Failed to reset password"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get reports with pagination",
                "operationId": "289342a77c685033524d27d2cb8bf192",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of records per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/deleterecord/{id}": {
            "delete": {
                "tags": [
                    "Reports"
                ],
                "summary": "Delete a report by ID",
                "operationId": "34041c95db1912787c5b746da04d58ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the report to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/count": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get the total count of reports",
                "operationId": "ff6aec0ff3db4242cb9f3ffb62af9577",
                "responses": {
                    "200": {
                        "description": "Returns the count of reports"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/reports/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export reports",
                "operationId": "5b74e8806b0c450925ec2fbcdc9a2cd6",
                "responses": {
                    "200": {
                        "description": "Exported reports"
                    },
                    "404": {
                        "description": "No data available"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getgroupreports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get group reports",
                "operationId": "eb0f3a37054c694e8258f811bd5d7418",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns group reports"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getcredibilityreports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get credibility reports",
                "operationId": "696d397a78b5acd36e1f19f6c3378b96",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter parameter",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns credibility reports"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/getemails": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get emails by group",
                "operationId": "111d1f5508c280a578e0aedccfa86204",
                "parameters": [
                    {
                        "name": "group",
                        "in": "query",
                        "description": "Group filter (default: A)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "A"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns emails"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/searchReports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Search reports",
                "operationId": "49382d4728736faba52f897fa4966279",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns search results"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}